+++ src/rogallo/commands/search.py
@@ -1,0 +1,15 @@
+"""Commands for searching for locations."""
+
+##############################################################################
+# Textual enhanced imports.
+from textual_enhanced.commands import Command
+
+
+##############################################################################
+class SearchHistory(Command):
+    """Search the history for a location."""
+
+    BINDING_KEY = "f6"
+
+
+### search.py ends here